+++ include/bedrock/nbt/short_tag.h
@@ -22,10 +22,12 @@
 class ShortTag : public Tag {
 public:
     explicit ShortTag(std::int16_t data = 0) : data(data) {}
+
     void write(IDataOutput &output) const override
     {
         output.writeShort(data);
     }
+
     Bedrock::Result<void> load(IDataInput &input) override
     {
         auto result = input.readShortResult();
@@ -35,23 +37,28 @@ public:
         }
         return nonstd::make_unexpected(result.error());
     }
+
     [[nodiscard]] std::string toString() const override
     {
         return std::to_string(data);
     }
+
     [[nodiscard]] Type getId() const override
     {
         return Type::Short;
     }
+
     [[nodiscard]] bool equals(const Tag &other) const override
     {
         return Tag::equals(other) && data == static_cast<const ShortTag &>(other).data;
     }
+
     [[nodiscard]] std::unique_ptr<Tag> copy() const override
     {
         return std::make_unique<ShortTag>(data);
     }
-    [[nodiscard]] std::uint64_t hash() const override
+
+    [[nodiscard]] std::size_t hash() const override
     {
         return data;
     }
