+++ include/bedrock/nbt/string_tag.h
@@ -14,7 +14,6 @@
 
 #pragma once
 
-#include <cstdint>
 #include <string>
 #include <utility>
 
@@ -24,10 +23,12 @@
 class StringTag : public Tag {
 public:
     explicit StringTag(std::string data = "") : data(std::move(data)) {}
+
     void write(IDataOutput &output) const override
     {
         output.writeString(data);
     }
+
     Bedrock::Result<void> load(IDataInput &input) override
     {
         auto result = input.readStringResult();
@@ -37,23 +38,28 @@ public:
         }
         return nonstd::make_unexpected(result.error());
     }
+
     [[nodiscard]] std::string toString() const override
     {
         return data;
     }
+
     [[nodiscard]] Type getId() const override
     {
         return Type::String;
     }
+
     [[nodiscard]] bool equals(const Tag &other) const override
     {
         return Tag::equals(other) && data == static_cast<const StringTag &>(other).data;
     }
+
     [[nodiscard]] std::unique_ptr<Tag> copy() const override
     {
         return std::make_unique<StringTag>(data);
     }
-    [[nodiscard]] std::uint64_t hash() const override
+
+    [[nodiscard]] std::size_t hash() const override
     {
         return std::hash<std::string>{}(data);
     }
